set4/pyproject.toml
[project]
name = "set4"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "emoji>=2.14.1",
    "inflect>=7.5.0",
    "pyfiglet>=1.0.3",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
]

[dependency-groups]
dev = [
    "pylint>=3.3.7",
]
